+++ src/classes/babies/BlackBaby.ts
@@ -4,8 +4,16 @@ import { pseudoRoomClearPostPEffectUpdateReordered } from "../features/PseudoRoo
 
 /** Curse Room doors in uncleared rooms. */
 export class BlackBaby extends Baby {
-  // We do not need to a `POST_NEW_ROOM_REORDERED` check because Curse Room doors do not need to be
-  // unlocked.
+  /**
+   * Flight means that they will be able to fly through the Curse Room doors, defeating the purpose
+   * of the mechanic.
+   */
+  override isValid(player: EntityPlayer): boolean {
+    return !player.CanFly;
+  }
+
+  // Unlike other babies that use the "pseudoRoomClear" feature, we do not need to a
+  // `POST_NEW_ROOM_REORDERED` check because Curse Room doors do not need to be unlocked.
 
   @CallbackCustom(ModCallbackCustom.POST_PEFFECT_UPDATE_REORDERED)
   postPEffectUpdateReordered(player: EntityPlayer): void {